backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
boto3==1.35.20
kubernetes==30.1.0
SQLAlchemy==2.0.35
alembic==1.13.2
apscheduler==3.10.4
pyyaml==6.0.2
PyGithub==2.4.0
httpx==0.27.2
dnspython==2.6.1
